images/projector-supra-vm/requirements.txt
aiohttp
jsonrpcserver
jsonrpcclient
pywinrm
impacket==0.11.0
psutil
paramiko
parallel-ssh
